backend/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
asyncpg
alembic
python-dotenv
